requirements_demos.txt
librosa>=0.9.1
numpy>=1.20.0
webrtcvad>=2.0.10
torch>=1.0.1
scipy>=1.2.1
typing
matplotlib>=3.0.0
sounddevice
tqdm
umap-learn
